application-generator/src/content/python/modules/emails/requirements.txt
blinker==1.5
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Mail==0.9.1
Flask-Moment==1.0.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==22.0
python-dotenv==0.21.0
Werkzeug==2.2.2
